Clean-up initialization of the NettackerModules class and moved socks_proxy to a separate file
